# docs/swagger.yaml
basePath: /
definitions:
  main.Recipe:
    properties:
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        items:
          type: string
        type: array
      name:
        type: string
      published_at:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: his is a sample recipes API. You can find out more about the API at
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Recipes API
  version: "1.0"
paths:
  /recipes:
    get:
      consumes:
      - application/json
      description: List all recipes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Recipe'
            type: array
      summary: List recipes
    post:
      consumes:
      - application/json
      description: Create a new recipe
      parameters:
      - description: Recipe
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/main.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Recipe'
      summary: Create a new recipe
  /recipes/{id}:
    put:
      consumes:
      - application/json
      description: Update a recipe
      parameters:
      - description: Recipe
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/main.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Recipe'
      summary: Update a recipe
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"